Moved sub-classes of "PropertyDescriptorGridEntry" and "PropertyStore" classes to their own files
