+++ src/System.Windows.Forms/src/System/Windows/Forms/PropertyGridInternal/PropertyDescriptorGridEntry.cs
@@ -15,7 +15,7 @@ using static Interop;
 
 namespace System.Windows.Forms.PropertyGridInternal
 {
-    internal class PropertyDescriptorGridEntry : GridEntry
+    internal partial  class PropertyDescriptorGridEntry : GridEntry
     {
         internal PropertyDescriptor _propertyInfo;
 
@@ -1151,295 +1151,6 @@ namespace System.Windows.Forms.PropertyGridInternal
         protected override GridEntryAccessibleObject GetAccessibilityObject()
         {
             return new PropertyDescriptorGridEntryAccessibleObject(this);
-        }
-
-        protected class PropertyDescriptorGridEntryAccessibleObject : GridEntryAccessibleObject
-        {
-            private readonly PropertyDescriptorGridEntry _owningPropertyDescriptorGridEntry;
-
-            public PropertyDescriptorGridEntryAccessibleObject(PropertyDescriptorGridEntry owner) : base(owner)
-            {
-                _owningPropertyDescriptorGridEntry = owner;
-            }
-
-            internal override bool IsIAccessibleExSupported() => true;
-
-            /// <summary>
-            ///  Returns the element in the specified direction.
-            /// </summary>
-            /// <param name="direction">Indicates the direction in which to navigate.</param>
-            /// <returns>Returns the element in the specified direction.</returns>
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                switch (direction)
-                {
-                    case UiaCore.NavigateDirection.NextSibling:
-                        var propertyGridViewAccessibleObject = (PropertyGridView.PropertyGridViewAccessibleObject)Parent;
-                        var propertyGridView = propertyGridViewAccessibleObject.Owner as PropertyGridView;
-                        bool currentGridEntryFound = false;
-                        return propertyGridViewAccessibleObject.GetNextGridEntry(_owningPropertyDescriptorGridEntry, propertyGridView.TopLevelGridEntries, out currentGridEntryFound);
-                    case UiaCore.NavigateDirection.PreviousSibling:
-                        propertyGridViewAccessibleObject = (PropertyGridView.PropertyGridViewAccessibleObject)Parent;
-                        propertyGridView = propertyGridViewAccessibleObject.Owner as PropertyGridView;
-                        currentGridEntryFound = false;
-                        return propertyGridViewAccessibleObject.GetPreviousGridEntry(_owningPropertyDescriptorGridEntry, propertyGridView.TopLevelGridEntries, out currentGridEntryFound);
-                    case UiaCore.NavigateDirection.FirstChild:
-                        return GetFirstChild();
-                    case UiaCore.NavigateDirection.LastChild:
-                        return GetLastChild();
-                }
-
-                return base.FragmentNavigate(direction);
-            }
-
-            private UiaCore.IRawElementProviderFragment GetFirstChild()
-            {
-                if (_owningPropertyDescriptorGridEntry is null)
-                {
-                    return null;
-                }
-
-                if (_owningPropertyDescriptorGridEntry.ChildCount > 0)
-                {
-                    return _owningPropertyDescriptorGridEntry.Children.GetEntry(0).AccessibilityObject;
-                }
-
-                var propertyGridView = GetPropertyGridView();
-                if (propertyGridView is null)
-                {
-                    return null;
-                }
-
-                var selectedGridEntry = propertyGridView.SelectedGridEntry;
-                if (_owningPropertyDescriptorGridEntry == selectedGridEntry)
-                {
-                    if (selectedGridEntry.Enumerable &&
-                        propertyGridView.DropDownVisible &&
-                        propertyGridView.DropDownControlHolder?.Component == propertyGridView.DropDownListBox)
-                    {
-                        return propertyGridView.DropDownListBoxAccessibleObject;
-                    }
-
-                    if (propertyGridView.DropDownVisible && propertyGridView.DropDownControlHolder is not null)
-                    {
-                        return propertyGridView.DropDownControlHolder.AccessibilityObject;
-                    }
-
-                    return propertyGridView.EditAccessibleObject;
-                }
-
-                return null;
-            }
-
-            private UiaCore.IRawElementProviderFragment GetLastChild()
-            {
-                if (_owningPropertyDescriptorGridEntry is null)
-                {
-                    return null;
-                }
-
-                if (_owningPropertyDescriptorGridEntry.ChildCount > 0)
-                {
-                    return _owningPropertyDescriptorGridEntry.Children
-                        .GetEntry(_owningPropertyDescriptorGridEntry.ChildCount - 1).AccessibilityObject;
-                }
-
-                var propertyGridView = GetPropertyGridView();
-                if (propertyGridView is null)
-                {
-                    return null;
-                }
-
-                var selectedGridEntry = propertyGridView.SelectedGridEntry;
-                if (_owningPropertyDescriptorGridEntry == selectedGridEntry)
-                {
-                    if (selectedGridEntry.Enumerable && propertyGridView.DropDownButton.Visible)
-                    {
-                        return propertyGridView.DropDownButton.AccessibilityObject;
-                    }
-
-                    return propertyGridView.EditAccessibleObject;
-                }
-
-                return null;
-            }
-
-            private PropertyGridView GetPropertyGridView()
-            {
-                var propertyGridViewAccessibleObject = Parent as PropertyGridView.PropertyGridViewAccessibleObject;
-                if (propertyGridViewAccessibleObject is null)
-                {
-                    return null;
-                }
-
-                return propertyGridViewAccessibleObject.Owner as PropertyGridView;
-            }
-
-            internal override bool IsPatternSupported(UiaCore.UIA patternId)
-            {
-                if (patternId == UiaCore.UIA.ValuePatternId ||
-                    (patternId == UiaCore.UIA.ExpandCollapsePatternId && owner.Enumerable))
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-
-            internal override void Expand()
-            {
-                if (ExpandCollapseState == UiaCore.ExpandCollapseState.Collapsed)
-                {
-                    ExpandOrCollapse();
-                }
-            }
-
-            internal override void Collapse()
-            {
-                if (ExpandCollapseState == UiaCore.ExpandCollapseState.Expanded)
-                {
-                    ExpandOrCollapse();
-                }
-            }
-
-            private void ExpandOrCollapse()
-            {
-                if (!GetPropertyGridView().IsHandleCreated)
-                {
-                    return;
-                }
-
-                var propertyGridView = GetPropertyGridView();
-                if (propertyGridView is null)
-                {
-                    return;
-                }
-
-                int row = propertyGridView.GetRowFromGridEntry(_owningPropertyDescriptorGridEntry);
-
-                if (row != -1)
-                {
-                    propertyGridView.PopupDialog(row);
-                }
-            }
-
-            internal override UiaCore.ExpandCollapseState ExpandCollapseState
-            {
-                get
-                {
-                    var propertyGridView = GetPropertyGridView();
-                    if (propertyGridView is null)
-                    {
-                        return UiaCore.ExpandCollapseState.Collapsed;
-                    }
-
-                    if (_owningPropertyDescriptorGridEntry == propertyGridView.SelectedGridEntry &&
-                        ((_owningPropertyDescriptorGridEntry is not null && _owningPropertyDescriptorGridEntry.InternalExpanded)
-                         || propertyGridView.DropDownVisible))
-                    {
-                        return UiaCore.ExpandCollapseState.Expanded;
-                    }
-
-                    return UiaCore.ExpandCollapseState.Collapsed;
-                }
-            }
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                if (propertyID == UiaCore.UIA.IsEnabledPropertyId)
-                {
-                    return !((PropertyDescriptorGridEntry)owner).IsPropertyReadOnly;
-                }
-                else if (propertyID == UiaCore.UIA.LegacyIAccessibleDefaultActionPropertyId)
-                {
-                    return string.Empty;
-                }
-                else if (propertyID == UiaCore.UIA.IsValuePatternAvailablePropertyId)
-                {
-                    return true;
-                }
-
-                return base.GetPropertyValue(propertyID);
-            }
-        }
-
-        /// <summary>
-        ///  The exception converter is a type converter that displays an exception to the user.
-        /// </summary>
-        private class ExceptionConverter : TypeConverter
-        {
-            /// <summary>
-            ///  Converts the given object to another type.  The most common types to convert
-            ///  are to and from a string object.  The default implementation will make a call
-            ///  to ToString on the object if the object is valid and if the destination
-            ///  type is string.  If this cannot convert to the destination type, this will
-            ///  throw a NotSupportedException.
-            /// </summary>
-            public override object ConvertTo(ITypeDescriptorContext context, CultureInfo culture, object value, Type destinationType)
-            {
-                if (destinationType == typeof(string))
-                {
-                    if (value is Exception ex)
-                    {
-                        if (ex.InnerException is not null)
-                        {
-                            ex = ex.InnerException;
-                        }
-
-                        return ex.Message;
-                    }
-
-                    return null;
-                }
-
-                throw GetConvertToException(value, destinationType);
-            }
-        }
-
-        /// <summary>
-        ///  The exception editor displays a message to the user.
-        /// </summary>
-        private class ExceptionEditor : UITypeEditor
-        {
-            public override object EditValue(ITypeDescriptorContext context, IServiceProvider provider, object value)
-            {
-                if (value is Exception except)
-                {
-                    IUIService uis = null;
-
-                    if (context is not null)
-                    {
-                        uis = (IUIService)context.GetService(typeof(IUIService));
-                    }
-
-                    if (uis is not null)
-                    {
-                        uis.ShowError(except);
-                    }
-                    else
-                    {
-                        string message = except.Message;
-                        if (message is null || message.Length == 0)
-                        {
-                            message = except.ToString();
-                        }
-
-                        RTLAwareMessageBox.Show(null, message, SR.PropertyGridExceptionInfo,
-                            MessageBoxButtons.OK, MessageBoxIcon.Error, MessageBoxDefaultButton.Button1, 0);
-                    }
-                }
-
-                return value;
-            }
-
-            /// <summary>
-            ///  Retrieves the editing style of the Edit method.  If the method
-            ///  is not supported, this will return None.
-            /// </summary>
-            public override UITypeEditorEditStyle GetEditStyle(ITypeDescriptorContext context)
-            {
-                return UITypeEditorEditStyle.Modal;
-            }
         }
     }
 }
